Move ReplacePseudoMemoryOpsPass to separate file
