+++ CMakeLists.txt
@@ -403,6 +403,9 @@ target_sources(${PROJECT_NAME}
         "common/version.h"
 )
 
+# ISPC passes
+add_subdirectory(src/opt)
+
 # To show stdlib.ispc in VS solution:
 if (WIN32)
     set_source_files_properties("${CMAKE_CURRENT_SOURCE_DIR}/stdlib.ispc" PROPERTIES HEADER_FILE_ONLY TRUE)
